Grid Clean-Up Consolidating all card containers
