+++ kolibri/plugins/learn/assets/src/views/content-cards/content-card.vue
