+++ kolibri/plugins/learn/assets/src/views/content-cards/content-carousel.vue
@@ -1,15 +1,13 @@
 <template>
 
   <div>
-    <content-grid :header="computedTitle" v-if="slicedContents.length">
-      <content-grid-item
+      <content-card
         v-for="content in slicedContents"
         :title="content.title"
         :thumbnail="content.thumbnail"
         :kind="content.kind"
         :progress="content.progress"
         :link="genContentLink(content.id)"/>
-    </content-grid>
 
     <div class="button-wrapper" v-if="contents.length > nCollapsed">
       <icon-button @click="toggle()" :text="less" v-if="expanded">
@@ -55,8 +53,7 @@
     },
     components: {
       'icon-button': require('kolibri.coreVue.components.iconButton'),
-      'content-grid-item': require('../content-grid/content-grid-item'),
-      'content-grid': require('../content-grid'),
+      'content-card': require('./content-card'),
     },
     data() {
       return {
