+++ kolibri/plugins/learn/assets/src/views/content-cards/content-grid.vue
@@ -1,0 +1,21 @@
+<template>
+
+  <div class="pure-g">
+    <slot/>
+  </div>
+
+</template>
+
+
+<script>
+
+  module.exports = {};
+
+</script>
+
+
+<style>
+
+  @require '~kolibri.styles.definitions'
+
+</style>
