+++ kolibri/plugins/learn/assets/src/views/content-cards/content-list/list-item/index.vue
@@ -39,7 +39,7 @@
 <style lang="stylus" scoped>
 
   @require '~kolibri.styles.definitions'
-  @require '../../learn.styl'
+  @require '../../../learn.styl'
 
   .list-item
     background-color: $core-bg-light
