+++ kolibri/plugins/learn/assets/src/views/content-cards/index.vue
@@ -1,0 +1,44 @@
+<template>
+
+  <div>
+    <div class="title">
+      <h2 v-if="title">{{title}}</h2>
+      <slot name="titlebox"/>
+    </div>
+    <component :contents="contents" :is="container">
+      <slot/>
+    </component>
+  </div>
+
+</template>
+
+
+<script>
+
+  module.exports = {
+    props: [
+      'title',
+      'container',
+      'contents',
+      'title',
+    ],
+    components: {
+      carousel: require('./content-carousel'),
+      grid: require('./content-grid'),
+      list: require('./content-list'),
+    },
+  };
+
+</script>
+
+
+<style lang="stylus" scoped>
+
+  .title
+    margin-top: 2em
+    margin-bottom: 1.4em
+
+  .title h2
+    display: inline
+
+</style>
